chore: refactored agents and toolsets internal abstractions.
