+++ docs/env-vars.md
@@ -1,19 +1,20 @@
 # ENVIRONMENT VARIABLES
 
 # General settings
-| Environment Variable           | Description                                 | Default Value          |
-|--------------------------------|---------------------------------------------|------------------------|
-| `AEGIS_CLI_FEATURE_AGENT`      | Set to `redhat` to use rh profile           | `public`               |
-| `AEGIS_LLM_HOST`               | Aegis LLM host                              | `localhost:11434`      |
-| `AEGIS_LLM_MODEL`              | Aegis LLM model                             | `llama3.2:latest`      |
-| `AEGIS_LLM_INPUT_TOKENS_WARN_THR` | Threshold for LLM input tokens to log a warning | 16384           |
-| `AEGIS_LLM_MAX_JOBS`           | Number of LLM prompts running concurrently  | 4                      |
-| `AEGIS_LLM_TIMEOUT_SECS`       | Timeout in seconds for a single LLM prompt  | 300                    |
-| `AEGIS_SAFETY_ENABLED`         | Enable separate model to check model safety | `false`                |
-| `AEGIS_SAFETY_LLM_HOST`        | Safety LLM host                             | `localhost:11434`      |
-| `AEGIS_SAFETY_LLM_MODEL`       | Safety LLM model                            | `granite3-guardian-2b` |
-| `AEGIS_SAFETY_OPENAPI_KEY`     | Safety openai key                           |                        |
-| `AEGIS_ML_CVE_DATA_DIR`        | Directory containing CVE training data      |                        |
+| Environment Variable           | Description                                     | Default Value          |
+|--------------------------------|-------------------------------------------------|------------------------|
+| `AEGIS_CLI_FEATURE_AGENT`      | Set to `redhat` to use rh profile               | `public`               |
+| `AEGIS_LLM_HOST`               | Aegis LLM host                                  | `localhost:11434`      |
+| `AEGIS_LLM_MODEL`              | Aegis LLM model                                 | `llama3.2:latest`      |
+| `AEGIS_LLM_INPUT_TOKENS_WARN_THR` | Threshold for LLM input tokens to log a warning | 16384                  |
+| `AEGIS_LLM_MAX_JOBS`           | Number of LLM prompts running concurrently      | 4                      |
+| `AEGIS_LLM_TIMEOUT_SECS`       | Timeout in seconds for a single LLM prompt      | 300                    |
+| `AEGIS_SAFETY_ENABLED`         | Enable separate model to check model safety     | `false`                |
+| `AEGIS_SAFETY_LLM_HOST`        | Safety LLM host                                 | `localhost:11434`      |
+| `AEGIS_SAFETY_LLM_MODEL`       | Safety LLM model                                | `granite3-guardian-2b` |
+| `AEGIS_SAFETY_OPENAPI_KEY`     | Safety openai key                               |                        |
+| `AEGIS_ML_CVE_DATA_DIR`        | Directory containing CVE training data          |                        |
+| `AEGIS_AGENT_MAX_RETRIES`        | Agent max retries                               | 5                      |
 
 
 # REST API settings
